refactor(components): [timeline] use JSX in Unit test (#8339)
